% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{check_vd}
\alias{check_vd}
\title{Check if visidata is installed}
\usage{
check_vd()
}
\value{
No return value. If visidata is not installed, the function stops execution.
}
\description{
This function checks if visidata is installed on the system.
If not, it stops the execution and prompts the user to install visidata.
}
