% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{vd}
\alias{vd}
\title{Open dataframe using visidata}
\usage{
vd(dat)
}
\arguments{
\item{dat}{The dataframe to open in visidata.}
}
\value{
The original dataframe, invisibly.
}
\description{
This function opens a dataframe in visidata. It first checks if jsonlite is installed,
and if so, it uses jsonlite to write the dataframe to a temporary json file.
If jsonlite is not installed, it writes the dataframe to a temporary csv file.
It then opens the temporary file in visidata.
}
